+++ src/main/java/kitchenpos/menu/infrastructure/JdbcTemplateMenuGroupDao.java
@@ -1,6 +1,7 @@
-package kitchenpos.menu.domain.dao;
+package kitchenpos.menu.infrastructure;
 
 import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.dao.MenuGroupDao;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
