+++ src/main/java/kitchenpos/order/infrastructure/JdbcTemplateOrderDao.java
@@ -1,7 +1,8 @@
-package kitchenpos.order.domain.dao;
+package kitchenpos.order.infrastructure;
 
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.domain.dao.OrderDao;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
