+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -12,9 +12,9 @@ import kitchenpos.NestedApplicationTest;
 import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.application.dto.MenuResponse;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.dao.JdbcTemplateMenuDao;
-import kitchenpos.menu.domain.dao.JdbcTemplateMenuGroupDao;
-import kitchenpos.product.domain.dao.JdbcTemplateProductDao;
+import kitchenpos.menu.infrastructure.JdbcTemplateMenuDao;
+import kitchenpos.menu.infrastructure.JdbcTemplateMenuGroupDao;
+import kitchenpos.product.infrastructure.JdbcTemplateProductDao;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.product.domain.Product;
