+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -12,8 +12,8 @@ import kitchenpos.table.application.dto.TableGroupResponse;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTables;
 import kitchenpos.table.domain.TableGroup;
-import kitchenpos.table.domain.dao.JdbcTemplateOrderTableDao;
-import kitchenpos.table.domain.dao.JdbcTemplateTableGroupDao;
+import kitchenpos.table.infrastructure.JdbcTemplateOrderTableDao;
+import kitchenpos.table.infrastructure.JdbcTemplateTableGroupDao;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
