+++ src/test/java/kitchenpos/infrastructure/JdbcTemplateTest.java
@@ -1,15 +1,15 @@
-package kitchenpos.dao;
+package kitchenpos.infrastructure;
 
 import javax.sql.DataSource;
-import kitchenpos.menu.domain.dao.JdbcTemplateMenuDao;
-import kitchenpos.menu.domain.dao.JdbcTemplateMenuGroupDao;
-import kitchenpos.menu.domain.dao.JdbcTemplateMenuProductDao;
-import kitchenpos.order.domain.dao.JdbcTemplateOrderDao;
-import kitchenpos.order.domain.dao.JdbcTemplateOrderLineItemDao;
-import kitchenpos.product.domain.dao.JdbcTemplateProductDao;
+import kitchenpos.menu.infrastructure.JdbcTemplateMenuDao;
+import kitchenpos.menu.infrastructure.JdbcTemplateMenuGroupDao;
+import kitchenpos.menu.infrastructure.JdbcTemplateMenuProductDao;
+import kitchenpos.order.infrastructure.JdbcTemplateOrderDao;
+import kitchenpos.order.infrastructure.JdbcTemplateOrderLineItemDao;
+import kitchenpos.product.infrastructure.JdbcTemplateProductDao;
 import kitchenpos.support.tool.DataBaseCleaner;
-import kitchenpos.table.domain.dao.JdbcTemplateOrderTableDao;
-import kitchenpos.table.domain.dao.JdbcTemplateTableGroupDao;
+import kitchenpos.table.infrastructure.JdbcTemplateOrderTableDao;
+import kitchenpos.table.infrastructure.JdbcTemplateTableGroupDao;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
